Changed mod structure because net.fabricmc was reserved, oopsie daisy
