+++ gradle.properties
@@ -8,9 +8,9 @@ org.gradle.jvmargs=-Xmx1G
 	loader_version=0.14.22
 
 # Mod Properties
-	modid=luck_of_the_draw
+	modid=luckofthedraw
 	mod_version = 0.0.1-TESTING-1.20.1
-	maven_group = net.fabricmc
+	maven_group = net.luckofthedraw
 	archives_base_name = Luck of the Draw
 
 # Dependencies
