+++ src/main/java/net/luckofthedraw/item/ModItems.java
@@ -1,10 +1,10 @@
-package net.fabricmc.luckofthedraw.item;
+package net.luckofthedraw.item;
 
 import net.fabricmc.fabric.api.item.v1.FabricItemSettings;
 import net.fabricmc.fabric.api.itemgroup.v1.FabricItemGroupEntries;
 import net.fabricmc.fabric.api.itemgroup.v1.ItemGroupEvents;
-import net.fabricmc.luckofthedraw.LuckOfTheDraw;
-import net.fabricmc.luckofthedraw.item.custom.*;
+import net.luckofthedraw.LuckOfTheDraw;
+import net.luckofthedraw.item.custom.*;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemGroups;
 import net.minecraft.registry.Registries;
