+++ src/main/java/net/luckofthedraw/item/custom/TheEmpressItem.java
@@ -1,4 +1,4 @@
-package net.fabricmc.luckofthedraw.item.custom;
+package net.luckofthedraw.item.custom;
 
 import net.minecraft.client.item.TooltipContext;
 import net.minecraft.item.ItemStack;
@@ -7,16 +7,15 @@ import net.minecraft.world.World;
 
 import java.util.List;
 
-public class TheEmperorItem extends MajorArcanaItem {
+public class TheEmpressItem extends MajorArcanaItem {
     // Item Settings
-    public TheEmperorItem(Settings majorArcanaItem) {
+    public TheEmpressItem(Settings majorArcanaItem) {
         super(majorArcanaItem,0);
     }
 
-
     // The Tooltip
     public void appendTooltip(ItemStack itemStack, World world, List<Text> tooltip, TooltipContext tooltipContext) {
-        tooltip.add(Text.translatable("item.luck_of_the_draw.the_emperor_tarot_card.tooltip_line1"));
-        tooltip.add(Text.translatable("item.luck_of_the_draw.the_emperor_tarot_card.tooltip_line2"));
+        tooltip.add(Text.translatable("item.luck_of_the_draw.the_empress_tarot_card.tooltip_line1"));
+        tooltip.add(Text.translatable("item.luck_of_the_draw.the_empress_tarot_card.tooltip_line2"));
     }
 }
